+++ src/app/game/page.tsx
@@ -1,12 +1,11 @@
 'use client';
 
-import MiniBoard from './MiniBoard';
+import MiniBoard from '@/components/MiniBoard';
 import { MiniBoardController } from '@/core/controllers/MiniBoardController';
 
 const miniBoardController = new MiniBoardController('a');
 
 export default function Game() {
-
   return (
     <div className="flex flex-col">
       <div>This is game page</div>
